requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cassandra-driver==3.29.2
certifi==2025.4.26
click==8.2.1
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.6.0
django-ninja==1.4.2
django-recaptcha==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
